+++ pkg/gateway/sys_darwin.go
@@ -3,10 +3,8 @@
 package gateway
 
 import (
-	"context"
 	"os/exec"
 
-	"github.com/cloudwego/netpoll"
 	"golang.org/x/sys/unix"
 )
 
@@ -26,20 +24,10 @@ func setTCPFastOpen(fd uintptr) error {
 	return nil
 }
 
-func setCloExec(fd int)  error {
+func setCloExec(fd int) error {
 	return nil
 }
 
 func setUserAndGroup(cmd *exec.Cmd, uid, gid uint32) {
 
-}
-
-
-
-func DisableGopool() error {
-	_ = netpoll.DisableGopool()
-	runTask = func(ctx context.Context, f func()) {
-		go f()
-	}
-	return nil
 }